% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularize.R
\name{pub_tabularize}
\alias{pub_tabularize}
\title{Tabularize chunks output}
\usage{
pub_tabularize(x)
}
\arguments{
\item{x}{input}
}
\description{
Tabularize chunks output
}
\examples{
# x <- system.file("examples/10_7554_eLife_03032.xml", 
#   package = "pubchunks")
# res <- pub_chunks(x, "elife", c("doi", "title", "keywords"))
# pub_tabularize(res)
}
