% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchunks-package.R
\docType{package}
\name{pubchunks-package}
\alias{pubchunks-package}
\alias{pubchunks}
\title{Get chunks of XML articles}
\author{
Scott Chamberlain \href{mailto:myrmecocystus@gmail.com}{myrmecocystus@gmail.com}
}
\keyword{package}
